requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
playwright
transformers
beautifulsoup4
lxml
boto3
Pillow
APScheduler 
openai
aiofiles 